Move VersionCompare struct to it's own module
